Clarify distinction between tasks and tries to run tasks
A task comes from (for example) an app invocation.

A task will be tried 0 or more times: 0 times if there is a memoized result,
more than 1 time if a try fails and a retry happens.

Most of this PR makes this distinction stronger in the monitoring system,
by splitting the 'task' table into two: 'task' and 'try'.