Tighten pytest KeyError test skipping
KeyError checking is only intended to happen for
configuration file loading.

On non-local tests, if the single configuration file
raises an exception, that exception can be propagated without
any special handling. So all of the KeyError handling
for non-local tests is removed by this commit.

On local tests, this commit adds some notes but does not
change behaviour for tests which raise a KeyError.

TODO: make local tests tighter as per added note