Formalized return from scale_in and scale_out. These were (somewhat) implemented by the two executors that actually did scaling (high_throughput and descendants, and low_latency), but not defined in the interface.
There was, however, a change to the return of the implementation of scale_in in both of the executors above. Before, it selected some block ids of blocks to kill, passed that to proivder.cancel() and return the result, which is a list of bools indicating which block ids were actually killed. However, this list would be nearly meaningless to a caller without access to the job ids whose removal was requested. Anyway, now it returns the list of block ids that were removed.