Merge two sections in handle_exec_update that have distinct ancestry
The first and second parts of handle_exec_update used to be in separate callbacks.
They were merged by #1431 to happen in the same callback, one after the other,
without any more fine grained merging of behaviour.

This commit merges some of the if statements to make retry behaviour clearer.

task record is deleted earlier now - as soon as a exception/result is set in app future, which has
moved, which means that subsequent code after this can't look up the task record. i've switch to
taking an explicit reference to the task record dictionary, which stops that happening mostly. but some calls appear to take a task_id rather than a task record (eg monitoring) and those will probably break now.

TODO: audit those calls which take a task_id in the exec callback code, and switch them to using
task record