Make dbm logging use same style as rest of parsl
practically this is so that I can access the logger without needing to pass around a dbm object,
which means I can still log even when that dbm object failed to be initialized.

this makes db manager logs appear in parsl log (even though they're in a different process) --
presumably because at process fork time, the global logging state is inherited. this might be
desirable, or it might be undesirable? I'm increasingly of the opinion that monitoring shouldn't
have its own separate log files if it is to be treated as a real component.