Remove unused resource info code: RESOURCE_INFO constant. unused task_id parameter to UDPRadio.sent. unused message type paramter to UDPRadio.send
Extra notes:

 commit 63b58e94aae23779a1966f9f757e08cadf92d23c Date:   Mon Feb 11 20:54:06 2019 -0600

renamed TASK_REG and TASK_INFO   to TASK_INFO, RESOURCE_INFO resp. but didn't replace the old TASK_INFO with new RESOURCE_INFO. maybe that's the introduction of this problem?

in the code prior to that commit, TASK_INFO was only used in the monitoring loop
and maybe should have been renamed to RESOURCE_INFO by 63b58...

Maybe make that change now, for potential simplification of code paths?

It's unclear to me how stuff ends up on the resource queue in the first place, rather than the priority queue?
In which case this tag is not necessary at all?

In which case, a further change would be to remove that tag entirely?