Replace executor __repr__ with executor label in periodic state report
Prior to this, I was seeing 38 lines of executor configuration sent to the
parsl log file every 5 seconds.

This commit adds python 3.5 compatible code to declare the type of label,
so that `make mypy` will pass. This code is a bit ugly but can be replaced
once we drop support for Python 3.5