Add support for Work Queue auto labeling resources (#1677)
* Add autolabel option to WQ executor

Uses the Resource Monitor to observe actual resource utilization
when running tasks. This allows for quickly dispatching multiple
tasks to large workers. Tasks that exceed initial resource limits
will be retried with bigger allocations. To make the best use of
this feature, tasks with significantly different resource
utilization should be in different categories.

* Adjust parameters for autolabeling

* Put each app in its own WQ category

* Add option to control WQ categories from app names

* Add config flag to tune category-steady-n-tasks in WQ executor

* Clean up autocategory check