wq excutor: treat parsl-default as any other category
When not using autocategory, the category of the task is set to
'parsl-default' before sending it to work queue. The work queue code
then can treat parsl-default as any other category. This simplifies the
logic a little bit, as the category of a task is set once and never
changed.