Wait for tasks to complete after each test in pytest (#1608)
* Wait for tasks to complete after each test in pytest

This will hang if tests are not properly completed by a test, rather
than those improperly completed tests falling off the end of the run.

At time of commit, that is happening with:
parsl/tests/test_python_apps/test_memoize_bad_id_for_memo.py

* Add a test for suspected bad behaviour around bad memoization and dependencies

Exception-at-app-submission doesn't work if the exception isn't thrown until dependencies complete

* Mark task as failed if launch() raises an exception rather than returning a Future

Prior to this PR, an exception raised by launch() propagates upwards, in at least two
different ways:

* If a task has no pending dependencies, then the exception propages to the user's code which
invoked the relevant app. An example of testing this can be seen in
parsl/tests/test_python_apps/test_memoize_bad_id_for_memo.py
which expects such an exception.

* If a task has pending dependencies, then the exception propagates into the dependency Future's
callback handler, and then out of the top of that into nowhere. The user is never informed. This
behaviour was not tested in the test suite.

In both cases, anything which waits on such tasks completing (for example, via AppFuture.result()
or downstream dependencies) will hang forever as the tasks never reach a final state.

This PR places error handling code in launch_if_ready so that if launch() raises an exception
rather than returning a Future, that exception is recorded and so the task is marked as failed.

The immediate problem that this fixes is when user provided memoization code raises an exception,
parsl hangs as above, rather than reporting a task failure. This blocks PRs #1608, #1614 from
being merged.

See discussion on PR #1275

* fix flake8

Co-authored-by: Yadu Nand Babuji <yadudoc1729@gmail.com>