Simplify user-facing function checkpointing behaviour
This PR makes function checkpointing be based only on the function name
and module name.

Prior to this PR, function checkpointing behaviour was complicated to
describe to users. For example:
 * some modifications to earlier parts of the source file would invalidate
   checkpoints unexpectedly, if the modification caused the line number of
   the checkpointed function definition to change.
 * other modifications, for example, to a function called by a checkpointed
   function, would not invalidate checkpoints, even though the result would
   now be different - because the checkpointing was based only on the
   definition of the function, not anything invoked by that function

With this PR, changing source code anywhere will not affect checkpointing
of functions, except when changing the name of the function.

This is discussed (along with alternatives) in GitHub issue #1591. Of all the
alternatives, the approach implemented in this PR feels to me like the
easiest to explain and understand.

Users can override function checkpoint identity using the existing per-type
singledispatch mechanism, to implement more complicated schemes, if/when they
need to customise this behaviour.

This commit also removes a test that checks that memoization *isn't* used
when function bodies differ, because that is now an incorrect assertion.