Force source=True when packing Python environments in WQ executor
The current worker wrapper script uses non-stdlib dependencies
(just ipyparallel) to deal with serialized functions. This poses
a problem for pack=True, since that dependency isn't in the body
of the function actually being sent. This is a quick fix to avoid
that code path.

For cases where we're providing the full Python environment, it
would be better if the worker wrapper didn't carry any non-stdlib
dependencies. This issue could alternatively be solved by
analyzing both the user function and the worker script.