port app page to new task/try schema
this removes the executor and submitted times from the table, because those
fields no longer exist at the task level. it might be better to put in some try summarisation here?

viz models.py is updated a bit, but not fully - it's an awkward mirror of the schema defined
in the main monitoring db_manager, using a different library. is there a better way to do this?
it's especially awkward as there are arbitrary SQL queries (including JOINs) in the viz code...