work queue executor: rework how file names are handled (#1705)
This pr is to solve issue #1228.

There was some name clashing when the executor was used across the different test sets. This pr changes the way the names of files that need to be staged in/out are handled. Now each tasks gets its own directory, and the remapping of .local_paths is done at the worker from a dictionary. Before, each name had to be tracked, and new names were generated on demand. However this logic was sometimes wrong, and it is much simpler to compute a single distinct name for the directory of the task.

To match this change:

The staging of stdout and stderr was simplified to match that of the rest of the files.
Files with url's but no .local_name's are not staged by work queue.
The wrapper that executes the task at the worker was modified to receive the dictionary of names as pickled file. With this, the wrapper is always called as: workqueue_worker map_file function_file result_file.
work queue code is tested both with and without shared filesystems. For shared filesystems only tests marked with issue363 are considered.