Added ability to capture errors during block launch that would otherwise propagate up the call stack and be discarded. These errors are then used to report the status of the respective blocks as FAILED.
This isn't quite right. The proper model when launching some async job is for the launch() method to either successfully launch the job such that a call to status() is valid or throw an exception, not fake a successful launch. However, until such time that the calling code properly handles such exceptions, this will have to do.