Merge two sections in handle_exec_update that have distinct ancestry (#1761)
The first and second parts of handle_exec_update used to be in separate callbacks,
They were merged by #1431 to happen in the same callback, one after the other,
without more fine grained merging of behaviour.

This commit merges some of the if statements to make retry behaviour clearer.
Previously the retry behaviour was split between the two sections described aboe.

This commit will peturb the order in which post-task completion events happening.
This means that task garbage collection can happen earlier on, and so looking up
a task_id in self.tasks cannot happen so much. _create_task_log_info is modified
to take a task record, rather than a task_id, to accomodate this.