Remove markup checkboxes that conflict with github's interpretation
Prior to this PR, the type of PR was presented as a series of checkboxes
with the intention that applicable ones are checked.

Github interprets checkboxes in PRs as todo lists, and renders a summary of them as "n of m
completed" progess bars on the PR summary page. This is wrong and confusing.

This PR moves formatting away from that style to prevent spurious github progress bars.