Add @join_apps
@join_apps allow a python app to launch more apps and then complete
only after the launched apps are completed, without occupying a
worker waiting.

These are described more in docs/userguide/joins.rst
which is added by this patch.

This PR adds a new state after the body of a python app completes,
but before the join future is complete, which is only entered
if an app is a join app.

This PR modifies the recursive fibonacci test case (which previously
blocked workers and so would deadlock easily) to use join apps.

This PR adds a field to the monitoring database to represent the
app to be joined on, which is similar to but not the same as a regular
dependency: a dependency indicates what must be completed for this
app to start, but joins field indicates what must be completed for
this app to complete.