Support for reporting asynchronous provider errors (#1783)
* Fix issues (#1706) with the single node launcher.
Added a debug flag, True by default, that, if set to False, suppresses stdout launcher messages. This is used for testing since it instructs the launcher to only produce messages from the launched command.

* Added exit code, stdout, and stderr fileds to JobStatus. These can be checked in order to get more detailed information from a completed task.

* Capture the exit code and streams of commands in the local provider.
Throw SubmitException when the submit() call fails before the status() call on the task id, should it exist, could function.

* Add test to check that both local and SSH channels work in the local provider and that both failing and succeeding tasks are properly reported as such.

* Also test with jobs_per_node > 1

* Import things from the right place

* Fixed job status comparison

* Added ability to capture errors during block launch that would otherwise propagate up the call stack and be discarded. These errors are then used to report the status of the respective blocks as FAILED.

This isn't quite right. The proper model when launching some async job is for the launch() method to either successfully launch the job such that a call to status() is valid or throw an exception, not fake a successful launch. However, until such time that the calling code properly handles such exceptions, this will have to do.

* Added an error handler which is launched by the status poller and detects (some) critical errors with executors.

* Make sure we also check for the actual process when getting the status. The previous solution ignored the possibility that the job could be killed.

* Import things from the right place

* Fixed job status comparison

* Added ability to capture errors during block launch that would otherwise propagate up the call stack and be discarded. These errors are then used to report the status of the respective blocks as FAILED.

This isn't quite right. The proper model when launching some async job is for the launch() method to either successfully launch the job such that a call to status() is valid or throw an exception, not fake a successful launch. However, until such time that the calling code properly handles such exceptions, this will have to do.

* Added an error handler which is launched by the status poller and detects (some) critical errors with executors.

* Add stdout and stderr info from status, now that they are available

* Make sure treshold is at least one, otherwise things will fail without actual failures

* Added tests to 1. validate basic htex and 2. check that a missing worker executable does lead to an error

* style fixes

* Updated doc stubs

* removed unused method

* Fix _read_summary, the method that reads an abridged stdout/stderr from a failed job to work properly with simulated job failures (which happen when the provider.submit() method throws an exception) and to correctly handle situations when no stdout/stderr is produced by a process.

Co-authored-by: Yadu Nand Babuji <yadudoc1729@gmail.com>