Clarify distinction between tasks and tries to run tasks (#1808)
* Clarify distinction between tasks and tries to run tasks

A task comes from (for example) an app invocation.

A task will be tried 0 or more times: 0 times if there is a memoized result,
more than 1 time if a try fails and a retry happens.

Most of this PR makes this distinction stronger in the monitoring system,
by splitting the 'task' table into two: 'task' and 'try'.

* Clarify distinction between tasks and tries to run tasks

A task comes from (for example) an app invocation.

A task will be tried 0 or more times: 0 times if there is a memoized result,
more than 1 time if a try fails and a retry happens.

Most of this PR makes this distinction stronger in the monitoring system,
by splitting the 'task' table into two: 'task' and 'try'.

* port app page to new task/try schema

this removes the executor and submitted times from the table, because those
fields no longer exist at the task level. it might be better to put in some try summarisation here?

viz models.py is updated a bit, but not fully - it's an awkward mirror of the schema defined
in the main monitoring db_manager, using a different library. is there a better way to do this?
it's especially awkward as there are arbitrary SQL queries (including JOINs) in the viz code...

* make tasks per app running plot use schema - should use combo of tasks and tries because it talks about
actually running stuff (which is what a try is) rather than about an app invocation (which is what
a task is)

This highlights a bug where everything is rounded to 1-second buckets and so if multiple apps run
in the same second, then the peak looks much higher. Now that tries are counted individually,
those false peaks are even higher sometimes.

* add fail_retryable to DAG plot

* remove spurious whitespace change

* add comment about try table

* Rerrange task/try timestamp columns

Tasks now have timestamps for:
  invocation
  return

Tries now have timestamps for:
  launch
  running
  return

* fix monitoring plots for recent db schema changes

* remove duplicate column name