I think there's a bug with inhibiting input staging, because input staging is inhibited for task that run on the data_manager executor, not arbitrary staging tasks which might run anywhere.
I think this might manifest on stageout tasks that run on arbitrary executors rather than on
data_manager, which is a case that is not covered in the default staging suite.

stageout tasks have an input set to be the file to be staged out.

So how about making a staging test for this with a toy staging provider?