Make monitoring more resilient to spurious network traffic
In practice on cori.nersc.gov on large runs, some unexpected network
traffic arrived at monitoring ports. This would sometimes crash
the monitoring system. This PR makes monitoring more likely to ignore
such traffic rather than crashing.

This PR adds some tests to generate spurious traffic.

Notes to reviewers:
I'm a little unclear about the correct way to report that
exceptions happened in the database loop: at the moment, this
ignores such exceptions and then reports a final exception
when monitoring finishes.

Another of my concerns is that there are other causes of exceptions,
not only spurious messages, and this may not be the right
handling for such messages. (for example, does a particular
socket closed exception result in the loop going into a 100%
CPU loop?)