tom has seen at scale the monitoring DB reporting "db locked" errors [A] and then [B] lots of "tried to update NNNN rows, only did MMMM" errors.
hypothesis:

i) we need to rety on db locked? [yes we do]

 at the sqlite3 commandline that needs to happen, at least, so
maybe also in code? this would only really show up if you're looking in the db logs, and probably
only if you're also running queries against the DB (or some other access) rather than letting the
db manager be the only user.

ii) if those locked errors are causing rows to not be INSERTed (eg for tasks) then that means that
row cannot be UPDATEd later, which might be then causing symptom B. That symptom probably means
that all updates in that batch are discarded. This could happen because of symptom A, or because
other malformed data could not be inserted into the database.

concerns:
 i) cross-DB portability (probably a lost cause without further abstraction, because
    that isn't really a thing SQL does)

 ii) what happens if I'm looping on some more genuine error that is always present and that
     causes some internal structures to overflow

For the purposes of lsst-parsl, this patch is allowed to be a bit of a messy hack, to try to understand
how this problem can be fixed.


Problems with this patch:
  I'm seeing database lock errors when running CREATE TABLE fairly often
  (but awkward enough to reproduce that I'm not confident in reproducing at the moment).
  This looks like sqlalchemy doing those ops somewhere that doesn't have the retry
  logic I added for doing inserts and updates. So this would manifest, i guess, if
  you were poking at the database before it had been properly initialised by the parsl
  monitoring code. I think that is probably a situation I care about.