Make monitoring switchable between UDPRadio and over-htex resource status delivery
TODO: needs to be switchable? At the moment, it is hard coded into the definition of an executor - which maybe makes sense when using HTEX, but not so much when parsl user wants to choose between mechanisms (eg UDP vs FS). That switching also needs to be per executor.

for lsst purposes this needs to scale up to the scale of
1000 nodes running many works (10s of) on each node.

so eg 50000 workers as a target?

50000 zmq TCP connections is A Lot

I'd like something that is not tied into the
htex codebase (which at the moment, monitoring is a bit,
and putting a messaging concentrator in there would
make it more tied in).

But a messaging concentrator per-node is what I need.
Perhaps I can re-use the existing htex queue, even,
using message types... "here's a message I want you to
forward to the messaging system in the interchange" ?

That would at least be a bit orthogonal? Or perhaps the
messaging engine (eg the UDPRadio bit) can be replaced
by an htex specific one. so you get UDPRadio in udp
circumstances, or HTEXRadio in the htex case, if you
so configure it? that would be an ok plugin thing
to try out that at least would allow switching
between UDP and HTEX while allowing very-HTEX specific
stuff to be implemented.

(and specificlaly, HTEX provides this kind of hierarchical
feature that parsl-core does not)

So additional communication this needs:

  i) monitoring worker wrapper: ability to communicate with htex worker enough to send monitoring messages
        - needs some "examine my context" that is probably provided by globals: if you're using "htex" mode monitoring, inspect a global that you expect to exist
        - and htex worker should set up that worker.
        - perhaps that also provides some scope for *inside apps* reporting state changes which is something i was interested in

  ii) interchange<->mangager<->htex worker: add in ability to send monitoring messages (perhaps as a broader "new tag to deliver")

  iii) interchange -> monitoring: deliver those messages to monitoring