Implement a better representation in monitoring for join apps
Before this patch, the joined future was added to the dependency list of the joinapp.

That doesn't properly capture joins, which is something a bit more like a
"contains" or "is composed of" relationship (but not entirely that).

This patch adds a new column into the joinapp to represent that.