Implement a better representation in monitoring for join apps
Before this patch, the joined future was added to the dependency list of the joinapp.

That doesn't properly capture joins, which is something a bit more like a
"contains" or "is composed of" relationship (but not entirely that).

This patch adds a new column into the try table to represent that.

Why the try table?
      a memoized join app won't launch any tasks/or do any joining
      a retried join app will launch a whole new set of tasks with a whole new final task

The tasks run to be joined on are part of the attempt to execute the app - so in that sense
they are sub-workflow-like.