Fix memo test that did not correctly enable caching
This test attempts to check that caching does not happen when
a function is redefined, but it was not enabling caching at all
because cache=True was in the wrong place.

This PR puts cache=True into the right place, and adds another
assertion that checks that caching *is* working when the function
is not redefined.