Split worker-side monitoring message code based on purpose
Prior to this patch, monitor() performed two functions, selected by a boolean
parameter:

i) It sends a task running message, which is used to populate the task_try_running
database field.

ii) It performs periodic monitoring of the process tree that is executing the
task.

These two code paths are fairly clearly separate, and do not need to both be in
monitor.

This PR is preparation for future PRs which allow worker-side monitoring to
report task start time without forking a resource monitoring process.