this is a prototype of using the filesystem to deliver monitoring messages
the file creation technique is similar to Maildir:
  create a file in tmp/ without any need to write atomically
  then move to new/ atomically

This means that the only messages appearing in new/ will be complete, and can be deleted by the receiver without conflicting with the sender.

I expect this will present interesting file system loads because of all the directory operations,
especially if everything is happening just inside a single directory. the load will perhaps be comparable to each job creating a few stdout/stderr files inside one directory. unless resource monitoring is on, in which case much higher load.

so maybe it will be slower but more reliable than UDP.

i expect that there will be more out-of-order message delivery than with UDP - and I expect monitoring will deal badly with that. (especially, the notion of "first" RESOURCE_INFO messages)

the intention is to use someone elses infrastructure to deliver messages rather than implementing a new htex-style comms network just for monitoring