Make block id type consistent in the status returned from executor status polling (#1952)
This PR refactors the key types in status returned from executor status polling method executor.status().

Before this PR, executor status returns a status dict, where the key of a normal running block is a provider job id, but the key of a failed block is the block id counter in executor internally. This inconsistency in block id type causes errors in monitoring (issue #1948 ). This PR addresses this inconsistency in block id type---for all blocks, their keys are the block id counter in executor internally.

This PR also renames external_block_id to block_id, and internal_block_id to job_id for ease to understand.

Note: it does not fix the issue in #1948 completely yet---upcoming PRs along with this PR will fix that.

* change status keys from internal to external

* renaming external internal block id

* mypy fixes

* mypy fixes

* update test to reflect get_job_ids change

* fix test

* fix mypy for job id

* fix mypy block_id type in make_status_dict

* docstring fix

* more executor status mypy type fixes

* mypy and docstring fixes

* flake8

* more block and job ids renaming

* scale in and out docstrings in executor base

* more rename

* make scale_in returns block_ids instead of job_ids

* remove block_id_type and fix monitoring bugs in #1948

* job_id can be null in monitoring db--fix issue #1948

* make docstring clear

* mypy fixes

* fix test_provider

Co-authored-by: Ben Clifford <benc@hawaga.org.uk>