More monitoring tests, using different configurations
I'd like to make this a bit more parameterised as there
are various monitoring configurations I want to try out.

My present model of copy-paste is not sustainable.

Some of this also expects radio mode to be defined per
executor class, not per executor instance. At least for
WQ, I now want to be able to switch that per instance.