tasks_per_node case in general strategy is a mess:
the ExtremeScale case will never fire in the current extremescale implementaiton,
because an extreme scale executor is also a high throughput executor, and so the
earlier htex case will fire.

It is possible that extreme scale scaling was broken because of this case. This
patch should not make it either better or worse, because it only eliminates dead
code.

when an executor is not an htex instance, no cases match, but no error is raised
here, and so tasks_per_node is never assigned. Later on (line 206) use of
tasks_per_node is an error.

this entire case is removed, and executor.workers_per_node is always used.