Pin sqlalchemy version due to sqlalchemy 1.4 release
SQLAlchemy 1.4.0 was released recently, and does not work
with the latest sqlalchemy utils:

SQLAlchemy-Utils                      0.36.8

This PR constrains the version of sqlalchemy to stay in
the 1.3.x range.

The import error is that _ColumnEntity has moved elsewhere in the
source tree of sqlalchemy, although this manifests in parsl monitoring
as a parsl error that monitoring dependencies are not installed, due to
assumptions about than an ImportError is because of uninstalled
dependecies.