fiddle with remote executor behaviour for better globus exceptions (eg #785)
dill claims it cannot serialize tracebacks, but that appears to be happening here... (or
at least, something is serializing tracebacks...)

The process worker pool can't pick tracebacks, though...
so it looks like something isn't being imported and registered there that is
imported and registered in end-worker code... ?

This PR moves to using the "ambient" executors serialization code, which might not
be correct.

Perhaps it should be explicit? And perhaps it should be the parsl serialize code rather
than dill explicitly?