Let RemoteExceptionWrappers serialize via executor serialization
Prior to this PR, RemoteExceptionWrappers explicitly serialised their wrapper
exception information using dill.

This serialization was done to give an object that could more easily be sent
over the network, as executor serializers could not serialize tracebacks at
the time.

However, the implementation would serialize even when this was not needed,
specifically when using the ThreadPoolExecutor.

This meant that some exceptions which did not serialize well (such as raised
by Globus, issue #785) but which were mostly only raised by tasks running in
the ThreadPoolExecutor (when using Globus file staging).

More recently, parsl has moved to expecting executors to be better able to
serialize things that can be serialised (see parsl.serialize) rather than
having ad-hoc serialization like this.

Fixes #2005