Allow join_apps to return multiple futures
Previously, to wait for multiple launched Futures inside an app,
it was necessary to make a separate `combine` app which waits for
them all the complete.

Use of this style of additional trivial app caused awkwardness in monitoring
and checkpointing: it often did not have a clear separate identity based on
parameters, which made it hard to distinguish "different" `combine` calls
from each other in the monitoring and checkpointing databases.

It was also boilerplate that seems nicer to not force on the user.