options to add more stuff into the task_hashsum based on: hashsum of dependencies
hashsum of dependencies (especially if done multi-level) is perhaps something
towards to dealing with "this thing X is re-run, but it returned the same
python-level result (eg None in the case of a bash app) - but despite that
I want everything downstream to re-run"? perhaps not. but the general idea of
more than just the python value from the previous step being part of the
"is this step already run?" is relevant I think - what else can we mix into
the hash to decide this?