/workflow/ffecef64-c6b0-472c-aa71-4363739d07e5/resource_usage fails:
...
  File "/home/benc/parsl/virtualenv-3.7/lib/python3.7/site-packages/flask/app.py", line 1936, in dispatch_request
    return self.view_functions[rule.endpoint](**req.view_args)
  File "/home/benc/parsl/src/parsl/parsl/monitoring/visualization/views.py", line 186, in workflow_resources
    worker_efficiency=worker_efficiency(df_task_tries, df_node),
  File "/home/benc/parsl/src/parsl/parsl/monitoring/visualization/plots/default/workflow_resource_plots.py", line 117, in worker_efficiency
    for j in range(int(row['epoch_time_running']), int(row['epoch_time_returned']) + 1):
ValueError: cannot convert float NaN to integer

with a test db from smoeone at lsst.

it looks like this is when there is a task that has started running but has not finished: this code assumes both of those fields are populated.

if the end field is not populated, assume the task is still runnning until the end time (which is roughly "now" or the end of the workflow)