Reverse polarity of mac safe queue imports to satisfy mypy
mypy determines the type of mpQueue by the type of its first usage.

Prior to this commit, that usage was importing MacSafeQueue.
That mean the second import did not type check: what it was importing
was not MacSafeQueue, but Queue.

This commit reverses that, so that mpQueue is now typed as a
Queue, and the subsequent import of MacSafeQueue is acceptable:
MacSafeQueue is a subclass of Queue.

The error message being fixed is this:
parsl/executors/high_throughput/process_worker_pool.py:28: error:
Incompatible import of "mpQueue" (imported name has type "Type[Queue[Any]]",
local name has type "Type[MacSafeQueue]")